Move metric key static creator to interface
