+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/metrics/MetricKeyUtil.java
@@ -4,8 +4,6 @@
 
 package net.sourceforge.pmd.lang.metrics;
 
-import java.util.Objects;
-
 import net.sourceforge.pmd.lang.ast.Node;
 
 /**
@@ -14,6 +12,7 @@ import net.sourceforge.pmd.lang.ast.Node;
  * @author Clément Fournier
  * @since 6.0.0
  */
+@Deprecated
 public final class MetricKeyUtil {
 
     private MetricKeyUtil() {
@@ -29,40 +28,11 @@ public final class MetricKeyUtil {
      * @param <T>    Type of node the metric can be computed on
      *
      * @return The metric key
+     *
+     * @deprecated Use {@link MetricKey#of(String, Metric)}
      */
-    // FUTURE Move that to the MetricKey interface once we upgrade the compiler
+    @Deprecated
     public static <T extends Node> MetricKey<T> of(final String name, final Metric<T> metric) {
-        return new MetricKey<T>() {
-            @Override
-            public String name() {
-                return name;
-            }
-
-
-            @Override
-            public Metric<T> getCalculator() {
-                return metric;
-            }
-
-
-            @Override
-            public boolean supports(T node) {
-                return metric.supports(node);
-            }
-
-
-            @Override
-            public boolean equals(Object obj) {
-                return obj != null && getClass() == obj.getClass()
-                    && Objects.equals(name(), ((MetricKey) obj).name())
-                    && Objects.equals(getCalculator(), ((MetricKey) obj).getCalculator());
-            }
-
-
-            @Override
-            public int hashCode() {
-                return (metric != null ? metric.hashCode() * 31 : 0) + (name != null ? name.hashCode() : 0);
-            }
-        };
+        return MetricKey.of(name, metric);
     }
 }
