+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/metrics/ResultOption.java
@@ -4,6 +4,8 @@
 
 package net.sourceforge.pmd.lang.metrics;
 
+import java.util.DoubleSummaryStatistics;
+
 /**
  * Options to calculate a result aggregated on the operations of a class. ResultOptions allow us to return the sum, or
  * average, or the highest value of the metric computed on the operations of a class. They help to reduce the need for
@@ -12,7 +14,11 @@ package net.sourceforge.pmd.lang.metrics;
  *
  * @author Clément Fournier
  * @since 5.8.0
+ * @deprecated Methods using this, like {@link MetricsUtil#computeAggregate(MetricKey, Iterable, ResultOption)}
+ *     are replaced with {@link DoubleSummaryStatistics}, use eg
+ *     {@link MetricsUtil#computeStatistics(MetricKey, Iterable)}
  */
+@Deprecated
 public enum ResultOption {
     /** Compute the sum on all operations. */
     SUM,
