+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/design/ExcessivePublicCountRule.java
@@ -4,6 +4,8 @@
 
 package net.sourceforge.pmd.lang.java.rule.design;
 
+import java.util.Objects;
+
 import net.sourceforge.pmd.lang.java.ast.ASTAnnotationTypeDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTAnyTypeBodyDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTAnyTypeDeclaration;
@@ -44,12 +46,12 @@ public class ExcessivePublicCountRule extends AbstractJavaCounterCheckRule<ASTAn
 
     @Override
     protected boolean isViolation(ASTAnyTypeDeclaration node, int reportLevel) {
-        long publicCount = node.getDeclarations()
+        long publicCount = node.getDeclarations().toStream()
                                .map(ASTAnyTypeBodyDeclaration::getDeclarationNode)
-                               .filterIs(AccessNode.class)
-                               .filter(AccessNode::isPublic)
+                               .filter(Objects::nonNull)
+                               .filter(it -> it instanceof AccessNode && ((AccessNode) it).isPublic())
                                // filter out constants
-                               .filter(it -> !(it instanceof ASTFieldDeclaration && it.isStatic() && it.isFinal()))
+                               .filter(it -> !(it instanceof ASTFieldDeclaration && ((ASTFieldDeclaration) it).isFinal() && ((ASTFieldDeclaration) it).isStatic()))
                                .count();
 
         return publicCount >= reportLevel;
